Finished separating text.rs system
