+++ amethyst_ui/src/text.rs
@@ -140,399 +140,136 @@ impl Component for TextEditing {
 }
 
 /// This system processes the underlying UI data as needed.
-pub struct UiKeyboardSystem<G> {
+/// TODO: Run after SelectionSystem
+pub struct TextEditingMouseSystem {
     /// A reader for winit events.
     reader: Option<ReaderId<Event>>,
     /// This is set to true while the left mouse button is pressed.
     left_mouse_button_pressed: bool,
     /// The screen coordinates of the mouse
     mouse_position: (f32, f32),
-    _phantom: PhantomData<G>,
 }
 
-impl<G> UiKeyboardSystem<G> {
+impl TextEditingMouseSystem {
     /// Creates a new instance of this system
     pub fn new() -> Self {
         Self {
             reader: None,
             left_mouse_button_pressed: false,
             mouse_position: (0., 0.),
-            _phantom: PhantomData,
         }
     }
 }
 
-impl<'a, G: Send+Sync+PartialEq+'static> System<'a> for UiKeyboardSystem<G> {
+impl<'a> System<'a> for TextEditingMouseSystem {
     type SystemData = (
-        Entities<'a>,
         WriteStorage<'a, UiText>,
         WriteStorage<'a, TextEditing>,
-        ReadStorage<'a, UiTransform>,
-        ReadStorage<'a, Selectable<G>>,
+        ReadStorage<'a, Selected>,
         Read<'a, EventChannel<Event>>,
-        Read<'a, Time>,
         ReadExpect<'a, ScreenDimensions>,
-        Read<'a, CachedSelectionOrder>
     );
 
     fn run(
         &mut self,
-        (entities, mut text, mut editable, transform, selectables, events, time, screen_dimensions, tab_order_cache): Self::SystemData,
+        (mut texts, mut text_editings, selecteds, events, screen_dimensions): Self::SystemData,
 ){
-        for text in (&mut text).join() {
+    	// Normalize text to ensure we can properly count the characters.
+    	// TODO: Possible improvement to be made if this can be moved only when inserting characters into ui text.
+        for text in (&mut texts).join() {
             if (*text.text).chars().any(is_combining_mark) {
                 let normalized = text.text.nfd().collect::<String>();
                 text.text = normalized;
             }
         }
 
-        {
-            let mut focused_text_edit = focused
-                .entity
+
+        // TODO: Link BlinkSystem
+        /*{
+            let mut text_edit = focused
                 .and_then(|entity| zip_options(text.get_mut(entity), editable.get_mut(entity)));
-            if let Some((ref mut _focused_text, ref mut focused_edit)) = focused_text_edit {
-                focused_edit.cursor_blink_timer += time.delta_real_seconds();
-                if focused_edit.cursor_blink_timer >= 1.0 / CURSOR_BLINK_RATE {
-                    focused_edit.cursor_blink_timer = 0.0;
+            if let Some((ref mut _text, ref mut text_editing)) = text_edit {
+                text_editing.cursor_blink_timer += time.delta_real_seconds();
+                if text_editing.cursor_blink_timer >= 1.0 / CURSOR_BLINK_RATE {
+                    text_editing.cursor_blink_timer = 0.0;
                 }
             }
-        }
-        for event in events.read(
-            self.reader
-                .as_mut()
-                .expect("`UiKeyboardSystem::setup` was not called before `UiKeyboardSystem::run`"),
-        ) {
-            // Process events for the whole UI.
-            match *event {
-                Event::WindowEvent {
-                    event: WindowEvent::CursorMoved { position, .. },
-                    ..
-                } => {
-                    let hidpi = screen_dimensions.hidpi_factor() as f32;
-                    self.mouse_position = (
-                        position.x as f32 * hidpi,
-                        (screen_dimensions.height() - position.y as f32) * hidpi,
-                    );
-                    if self.left_mouse_button_pressed {
-                        let mut focused_text_edit = focused.entity.and_then(|entity| {
-                            zip_options(text.get_mut(entity), editable.get_mut(entity))
-                        });
-                        if let Some((ref mut focused_text, ref mut focused_edit)) =
-                            focused_text_edit
-                        {
+        }*/
+
+        // Process only if an editable text is selected.
+        if let Some((ref mut text, ref mut text_editing, _)) = (&mut texts, &mut text_editings, &selecteds).join().next(){
+	        for event in events.read(
+	            self.reader
+	                .as_mut()
+	                .expect("`UiKeyboardSystem::setup` was not called before `UiKeyboardSystem::run`"),
+	        ) {
+	            // Process events for the whole UI.
+	            match *event {
+	                Event::WindowEvent {
+	                    event: WindowEvent::CursorMoved { position, .. },
+	                    ..
+	                } => {
+	                    let hidpi = screen_dimensions.hidpi_factor() as f32;
+	                    self.mouse_position = (
+	                        position.x as f32 * hidpi,
+	                        (screen_dimensions.height() - position.y as f32) * hidpi,
+	                    );
+	                    if self.left_mouse_button_pressed {
                             let (mouse_x, mouse_y) = self.mouse_position;
-                            focused_edit.highlight_vector = closest_glyph_index_to_mouse(
+                            text_editing.highlight_vector = closest_glyph_index_to_mouse(
                                 mouse_x,
                                 mouse_y,
-                                focused_text.cached_glyphs.iter(),
-                            ) - focused_edit.cursor_position;
+                                text.cached_glyphs.iter(),
+                            ) - text_editing.cursor_position;
                             // The end of the text, while not a glyph, is still something
                             // you'll likely want to click your cursor to, so if the cursor is
                             // near the end of the text, check if we should put it at the end
                             // of the text.
-                            if should_advance_to_end(mouse_x, focused_edit, focused_text) {
-                                focused_edit.highlight_vector += 1;
+                            if should_advance_to_end(mouse_x, text_editing, text) {
+                                text_editing.highlight_vector += 1;
                             }
-                        }
-                    }
-                }
-                Event::WindowEvent {
-                    event:
-                        WindowEvent::MouseInput {
-                            button: MouseButton::Left,
-                            state,
-                            ..
-                        },
-                    ..
-                } => {
-                    match state {
-                        ElementState::Pressed => {
-                            self.left_mouse_button_pressed = true;
+	                    }
+	                }
+	                Event::WindowEvent {
+	                    event:
+	                        WindowEvent::MouseInput {
+	                            button: MouseButton::Left,
+	                            state,
+	                            ..
+	                        },
+	                    ..
+	                } => {
+	                    match state {
+	                        ElementState::Pressed => {
+	                            self.left_mouse_button_pressed = true;
 
-                            focused.entity = (&*entities, &transform)
-                                .join()
-                                .filter(|&(_, t)| {
-                                    t.pixel_x - t.width / 2.0 <= self.mouse_position.0
-                                        && t.pixel_x + t.width / 2.0 >= self.mouse_position.0
-                                        && t.pixel_y - t.height / 2.0 <= self.mouse_position.1
-                                        && t.pixel_y + t.height / 2.0 >= self.mouse_position.1
-                                }).map(|(e, t)| (e, t.global_z))
-                                // In instances of ambiguity we want to select the element with the
-                                // highest Z order, so we need to find the highest Z order value among
-                                // eligible elements.
-                                .max_by(|(_, z1), (_, z2)| z1.partial_cmp(z2).expect("Z was NaN"))
-                                .map(|(e, _)| e);
-                            // If we focused an editable text field be sure to position the cursor
-                            // in it.
-                            let mut focused_text_edit = focused.entity.and_then(|entity| {
-                                zip_options(text.get_mut(entity), editable.get_mut(entity))
-                            });
-                            if let Some((ref mut focused_text, ref mut focused_edit)) =
-                                focused_text_edit
-                            {
+	                            // If we focused an editable text field be sure to position the cursor
+	                            // in it.
                                 let (mouse_x, mouse_y) = self.mouse_position;
-                                focused_edit.highlight_vector = 0;
-                                focused_edit.cursor_position = closest_glyph_index_to_mouse(
+                                text_editing.highlight_vector = 0;
+                                text_editing.cursor_position = closest_glyph_index_to_mouse(
                                     mouse_x,
                                     mouse_y,
-                                    focused_text.cached_glyphs.iter(),
+                                    text.cached_glyphs.iter(),
                                 );
 
                                 // The end of the text, while not a glyph, is still something
                                 // you'll likely want to click your cursor to, so if the cursor is
                                 // near the end of the text, check if we should put it at the end
                                 // of the text.
-                                if should_advance_to_end(mouse_x, focused_edit, focused_text) {
-                                    focused_edit.cursor_position += 1;
+                                if should_advance_to_end(mouse_x, text_editing, text) {
+                                    text_editing.cursor_position += 1;
                                 }
-                            }
-                        }
-                        ElementState::Released => {
-                            self.left_mouse_button_pressed = false;
-                        }
-                    }
-                }
-                _ => {}
-            }
-            let mut focused_text_edit = focused
-                .entity
-                .and_then(|entity| zip_options(text.get_mut(entity), editable.get_mut(entity)));
-            // Process events for the focused text element
-            if let Some((ref mut focused_text, ref mut focused_edit)) = focused_text_edit {
-                match *event {
-                    Event::WindowEvent {
-                        event: WindowEvent::ReceivedCharacter(input),
-                        ..
-                    } => {
-                        // Ignore obsolete control characters, and tab characters we can't render
-                        // properly anyways.  Also ignore newline characters since we don't
-                        // support multi-line text at the moment.
-                        if input < '\u{20}' {
-                            continue;
-                        }
-                        // Ignore delete character too
-                        else if input == '\u{7F}' {
-                            continue;
-                        }
-                        // Unicode reserves some characters for "private use".  Systems emit
-                        // these for no clear reason, so we're just going to ignore all of them.
-                        else if input >= '\u{E000}' && input <= '\u{F8FF}' {
-                            continue;
-                        } else if input >= '\u{F0000}' && input <= '\u{FFFFF}' {
-                            continue;
-                        } else if input >= '\u{100000}' && input <= '\u{10FFFF}' {
-                            continue;
-                        }
-                        focused_edit.cursor_blink_timer = 0.0;
-                        delete_highlighted(focused_edit, focused_text);
-                        let start_byte = focused_text
-                            .text
-                            .grapheme_indices(true)
-                            .nth(focused_edit.cursor_position as usize)
-                            .map(|i| i.0)
-                            .unwrap_or_else(|| {
-                                // We are either in a 0 length string, or at the end of a string
-                                // This line returns the correct byte index for both.
-                                focused_text.text.len()
-                            });
-                        if focused_text.text.graphemes(true).count() < focused_edit.max_length {
-                            focused_text.text.insert(start_byte, input);
-                            focused_edit.cursor_position += 1;
-                        }
-                    }
-                    Event::WindowEvent {
-                        event:
-                            WindowEvent::KeyboardInput {
-                                input:
-                                    KeyboardInput {
-                                        state: ElementState::Pressed,
-                                        virtual_keycode: Some(v_keycode),
-                                        modifiers,
-                                        ..
-                                    },
-                                ..
-                            },
-                        ..
-                    } => match v_keycode {
-                        VirtualKeyCode::Home | VirtualKeyCode::Up => {
-                            focused_edit.highlight_vector = if modifiers.shift {
-                                focused_edit.cursor_position
-                            } else {
-                                0
-                            };
-                            focused_edit.cursor_position = 0;
-                            focused_edit.cursor_blink_timer = 0.0;
-                        }
-                        VirtualKeyCode::End | VirtualKeyCode::Down => {
-                            let glyph_len = focused_text.text.graphemes(true).count() as isize;
-                            focused_edit.highlight_vector = if modifiers.shift {
-                                focused_edit.cursor_position - glyph_len
-                            } else {
-                                0
-                            };
-                            focused_edit.cursor_position = glyph_len;
-                            focused_edit.cursor_blink_timer = 0.0;
-                        }
-                        VirtualKeyCode::Back => {
-                            if !delete_highlighted(focused_edit, focused_text) {
-                                if focused_edit.cursor_position > 0 {
-                                    if let Some((byte, len)) = focused_text
-                                        .text
-                                        .grapheme_indices(true)
-                                        .nth(focused_edit.cursor_position as usize - 1)
-                                        .map(|i| (i.0, i.1.len()))
-                                    {
-                                        focused_text.text.drain(byte..(byte + len));
-                                        focused_edit.cursor_position -= 1;
-                                    }
-                                }
-                            }
-                        }
-                        VirtualKeyCode::Delete => {
-                            if !delete_highlighted(focused_edit, focused_text) {
-                                if let Some((start_byte, start_glyph_len)) = focused_text
-                                    .text
-                                    .grapheme_indices(true)
-                                    .nth(focused_edit.cursor_position as usize)
-                                    .map(|i| (i.0, i.1.len()))
-                                {
-                                    focused_edit.cursor_blink_timer = 0.0;
-                                    focused_text
-                                        .text
-                                        .drain(start_byte..(start_byte + start_glyph_len));
-                                }
-                            }
-                        }
-                        VirtualKeyCode::Left => {
-                            if focused_edit.highlight_vector == 0 || modifiers.shift {
-                                if focused_edit.cursor_position > 0 {
-                                    let delta = if ctrl_or_cmd(&modifiers) {
-                                        let mut graphemes = 0;
-                                        for word in focused_text.text.split_word_bounds() {
-                                            let word_graphemes =
-                                                word.graphemes(true).count() as isize;
-                                            if graphemes + word_graphemes
-                                                >= focused_edit.cursor_position
-                                            {
-                                                break;
-                                            }
-                                            graphemes += word_graphemes;
-                                        }
-                                        focused_edit.cursor_position - graphemes
-                                    } else {
-                                        1
-                                    };
-                                    focused_edit.cursor_position -= delta;
-                                    if modifiers.shift {
-                                        focused_edit.highlight_vector += delta;
-                                    }
-                                    focused_edit.cursor_blink_timer = 0.0;
-                                }
-                            } else {
-                                focused_edit.cursor_position = focused_edit.cursor_position.min(
-                                    focused_edit.cursor_position + focused_edit.highlight_vector,
-                                );
-                                focused_edit.highlight_vector = 0;
-                            }
-                        }
-                        VirtualKeyCode::Right => {
-                            if focused_edit.highlight_vector == 0 || modifiers.shift {
-                                let glyph_len = focused_text.text.graphemes(true).count();
-                                if (focused_edit.cursor_position as usize) < glyph_len {
-                                    let delta = if ctrl_or_cmd(&modifiers) {
-                                        let mut graphemes = 0;
-                                        for word in focused_text.text.split_word_bounds() {
-                                            graphemes += word.graphemes(true).count() as isize;
-                                            if graphemes > focused_edit.cursor_position {
-                                                break;
-                                            }
-                                        }
-                                        graphemes - focused_edit.cursor_position
-                                    } else {
-                                        1
-                                    };
-                                    focused_edit.cursor_position += delta;
-                                    if modifiers.shift {
-                                        focused_edit.highlight_vector -= delta;
-                                    }
-                                    focused_edit.cursor_blink_timer = 0.0;
-                                }
-                            } else {
-                                focused_edit.cursor_position = focused_edit.cursor_position.max(
-                                    focused_edit.cursor_position + focused_edit.highlight_vector,
-                                );
-                                focused_edit.highlight_vector = 0;
-                            }
-                        }
-                        VirtualKeyCode::A => {
-                            if ctrl_or_cmd(&modifiers) {
-                                let glyph_len = focused_text.text.graphemes(true).count() as isize;
-                                focused_edit.cursor_position = glyph_len;
-                                focused_edit.highlight_vector = -glyph_len;
-                            }
-                        }
-                        VirtualKeyCode::X => {
-                            if ctrl_or_cmd(&modifiers) {
-                                let new_clip = extract_highlighted(focused_edit, focused_text);
-                                if !new_clip.is_empty() {
-                                    if let Err(e) = ClipboardProvider::new().and_then(
-                                        |mut ctx: ClipboardContext| ctx.set_contents(new_clip),
-                                    ) {
-                                        error!("Error occured when cutting to clipboard: {:?}", e);
-                                    }
-                                }
-                            }
-                        }
-                        VirtualKeyCode::C => {
-                            if ctrl_or_cmd(&modifiers) {
-                                let new_clip = read_highlighted(focused_edit, focused_text);
-                                if !new_clip.is_empty() {
-                                    if let Err(e) = ClipboardProvider::new().and_then(
-                                        |mut ctx: ClipboardContext| {
-                                            ctx.set_contents(new_clip.to_owned())
-                                        },
-                                    ) {
-                                        error!("Error occured when copying to clipboard: {:?}", e);
-                                    }
-                                }
-                            }
-                        }
-                        VirtualKeyCode::V => {
-                            if ctrl_or_cmd(&modifiers) {
-                                delete_highlighted(focused_edit, focused_text);
-
-                                match ClipboardProvider::new()
-                                    .and_then(|mut ctx: ClipboardContext| ctx.get_contents())
-                                {
-                                    Ok(contents) => {
-                                        let index = cursor_byte_index(focused_edit, focused_text);
-                                        let empty_space =
-                                            focused_edit.max_length
-                                                - focused_text.text.graphemes(true).count();
-                                        let contents = contents
-                                            .graphemes(true)
-                                            .take(empty_space)
-                                            .fold(String::new(), |mut init, new| {
-                                                init.push_str(new);
-                                                init
-                                            });
-                                        focused_text.text.insert_str(index, &contents);
-                                        focused_edit.cursor_position +=
-                                            contents.graphemes(true).count() as isize;
-                                    }
-                                    Err(e) => error!(
-                                        "Error occured when pasting contents of clipboard: {:?}",
-                                        e
-                                    ),
-                                }
-                            }
-                        }
-                        _ => {}
-                    },
-                    _ => {}
-                }
-            }
-        }
+	                        }
+	                        ElementState::Released => {
+	                            self.left_mouse_button_pressed = false;
+	                        }
+	                    }
+	                }
+	                _ => {}
+	            }
+	        }
+    	}
     }
 
     fn setup(&mut self, res: &mut Resources) {
@@ -544,13 +281,13 @@ impl<'a, G: Send+Sync+PartialEq+'static> System<'a> for UiKeyboardSystem<G> {
 
 fn should_advance_to_end(
     mouse_x: f32,
-    focused_edit: &mut TextEditing,
-    focused_text: &mut UiText,
+    text_editing: &mut TextEditing,
+    text: &mut UiText,
 ) -> bool {
-    let cursor_pos = focused_edit.cursor_position + focused_edit.highlight_vector;
-    let len = focused_text.cached_glyphs.len() as isize;
+    let cursor_pos = text_editing.cursor_position + text_editing.highlight_vector;
+    let len = text.cached_glyphs.len() as isize;
     if cursor_pos + 1 == len {
-        if let Some(last_glyph) = focused_text.cached_glyphs.last() {
+        if let Some(last_glyph) = text.cached_glyphs.last() {
             let last_glyph_x = last_glyph.position().x;
             let advance_width = last_glyph.unpositioned().h_metrics().advance_width;
             if mouse_x - last_glyph_x > advance_width / 2.0 {
@@ -560,15 +297,6 @@ fn should_advance_to_end(
     }
 
     false
-}
-
-fn zip_options<T1, T2>(o1: Option<T1>, o2: Option<T2>) -> Option<(T1, T2)> {
-    if let Some(o1) = o1 {
-        if let Some(o2) = o2 {
-            return Some((o1, o2));
-        }
-    }
-    None
 }
 
 fn closest_glyph_index_to_mouse<'a, 'b: 'a, I>(mouse_x: f32, mouse_y: f32, i: I) -> isize
@@ -586,65 +314,3 @@ where
         .unwrap_or(0) as isize
 }
 
-/// Returns if the command key is down on OSX, and the CTRL key for everything else.
-fn ctrl_or_cmd(modifiers: &ModifiersState) -> bool {
-    (cfg!(target_os = "macos") && modifiers.logo)
-        || (cfg!(not(target_os = "macos")) && modifiers.ctrl)
-}
-
-fn read_highlighted<'a>(edit: &TextEditing, text: &'a UiText) -> &'a str {
-    let range = highlighted_bytes(edit, text);
-    &text.text[range]
-}
-
-/// Removes the highlighted text and returns it in a String.
-fn extract_highlighted(edit: &mut TextEditing, text: &mut UiText) -> String {
-    let range = highlighted_bytes(edit, text);
-    edit.cursor_position = range.start as isize;
-    edit.highlight_vector = 0;
-    text.text.drain(range).collect::<String>()
-}
-
-/// Removes the highlighted text and returns true if anything was deleted..
-fn delete_highlighted(edit: &mut TextEditing, text: &mut UiText) -> bool {
-    if edit.highlight_vector != 0 {
-        let range = highlighted_bytes(edit, text);
-        edit.cursor_position = range.start as isize;
-        edit.highlight_vector = 0;
-        text.text.drain(range);
-        return true;
-    }
-    false
-}
-
-// Gets the byte index of the cursor.
-fn cursor_byte_index(edit: &TextEditing, text: &UiText) -> usize {
-    text.text
-        .grapheme_indices(true)
-        .nth(edit.cursor_position as usize)
-        .map(|i| i.0)
-        .unwrap_or_else(|| text.text.len())
-}
-
-/// Returns the byte indices that are highlighted in the string.
-fn highlighted_bytes(edit: &TextEditing, text: &UiText) -> Range<usize> {
-    let start = edit
-        .cursor_position
-        .min(edit.cursor_position + edit.highlight_vector) as usize;
-    let end = edit
-        .cursor_position
-        .max(edit.cursor_position + edit.highlight_vector) as usize;
-    let start_byte = text
-        .text
-        .grapheme_indices(true)
-        .nth(start)
-        .map(|i| i.0)
-        .unwrap_or_else(|| text.text.len());
-    let end_byte = text
-        .text
-        .grapheme_indices(true)
-        .nth(end)
-        .map(|i| i.0)
-        .unwrap_or_else(|| text.text.len());
-    start_byte..end_byte
-}
